# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Task Management RESTful API Backend with authentication, workspaces, projects, tasks, statuses, and labels management.",
        "title": "GoTask API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/invoices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all invoices for the current workspace.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "List invoices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch invoices",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new invoice with auto-generated invoice number (INV-YYYY-XXX).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Create an invoice",
                "parameters": [
                    {
                        "description": "Invoice payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invoices.CreateInvoiceRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Invoice created successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create invoice",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/invoices/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single invoice by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Get an invoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Invoice not found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently delete an invoice.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Delete an invoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Invoice not found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an invoice. Setting status to \"paid\" triggers revenue sync to client.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Update an invoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Invoice update payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invoices.UpdateInvoiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice updated successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Invoice not found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/invoices/{id}/mark-paid": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks an invoice as paid and syncs revenue to client atomically.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Mark invoice as paid",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Mark paid payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invoices.MarkPaidRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice marked as paid",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error or invoice cancelled",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Invoice not found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "invoices.CreateInvoiceRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 500000
                },
                "client_id": {
                    "type": "integer"
                },
                "discount": {
                    "type": "number",
                    "example": 0
                },
                "due_date": {
                    "type": "string",
                    "example": "2026-06-30T00:00:00Z"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoices.InvoiceItemRequest"
                    }
                },
                "notes": {
                    "type": "string",
                    "example": "Payment due within 30 days"
                },
                "project_id": {
                    "type": "integer"
                },
                "tax": {
                    "type": "number",
                    "example": 50000
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "invoices.InvoiceItemRequest": {
            "type": "object",
            "required": [
                "description",
                "quantity",
                "total",
                "unit_price"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Web Development"
                },
                "quantity": {
                    "type": "number",
                    "example": 1
                },
                "total": {
                    "type": "number",
                    "example": 500000
                },
                "unit_price": {
                    "type": "number",
                    "example": 500000
                }
            }
        },
        "invoices.MarkPaidRequest": {
            "type": "object",
            "required": [
                "amount_paid"
            ],
            "properties": {
                "amount_paid": {
                    "type": "number",
                    "minimum": 0,
                    "example": 500000
                },
                "paid_at": {
                    "type": "string",
                    "example": "2026-05-20T10:00:00Z"
                }
            }
        },
        "invoices.UpdateInvoiceRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 500000
                },
                "client_id": {
                    "type": "integer"
                },
                "discount": {
                    "type": "number",
                    "example": 0
                },
                "due_date": {
                    "type": "string",
                    "example": "2026-06-30T00:00:00Z"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoices.InvoiceItemRequest"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "project_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string",
                    "example": "paid"
                },
                "tax": {
                    "type": "number",
                    "example": 50000
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "tier_info": {
                    "$ref": "#/definitions/utils.TierInfo"
                }
            }
        },
        "utils.TierInfo": {
            "type": "object",
            "properties": {
                "days_remaining": {
                    "type": "integer"
                },
                "expires_at": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "tier": {
                    "type": "string"
                },
                "warning": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Format: 'Bearer {token}'",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}